requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Django==5.1.1
django-allauth==64.2.1
idna==3.8
jwt==1.3.1
pillow==10.4.0
pycparser==2.22
PyJWT==1.7.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
svgwrite==1.4.3
Tree==0.2.4
typing_extensions==4.12.2
urllib3==2.2.2
